# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sentiment Analysis API",
    "description": "Entraînement et prédiction pour l'analyse de sentiments FR/EN",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/train/intent": {
      "post": {
        "tags": [
          "Intent Training v1"
        ],
        "summary": "Start Intent Training",
        "description": "Lance l'entraînement d'intention (job kind=\"intent\", 202).",
        "operationId": "start_intent_training_api_v1_train_intent_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntentTrainRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainJob"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/train/intent/status/{job_id}": {
      "get": {
        "tags": [
          "Intent Training v1"
        ],
        "summary": "Get Intent Training Status",
        "description": "Statut d'un job d'intention (404 si inconnu — use case partagé).",
        "operationId": "get_intent_training_status_api_v1_train_intent_status__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainJob"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/train/intent/cancel/{job_id}": {
      "post": {
        "tags": [
          "Intent Training v1"
        ],
        "summary": "Cancel Intent Training Endpoint",
        "description": "Annule un job d'intention (404 si inconnu).",
        "operationId": "cancel_intent_training_endpoint_api_v1_train_intent_cancel__job_id__post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainJob"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/train/intent/jobs": {
      "get": {
        "tags": [
          "Intent Training v1"
        ],
        "summary": "List Intent Training Jobs Endpoint",
        "description": "Liste paginée des jobs d'intention UNIQUEMENT (filtre kind=\"intent\").",
        "operationId": "list_intent_training_jobs_endpoint_api_v1_train_intent_jobs_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/JobStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrer par status : pending, running, completed, failed, cancelled",
              "title": "Status"
            },
            "description": "Filtrer par status : pending, running, completed, failed, cancelled"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Nombre max de résultats",
              "default": 100,
              "title": "Limit"
            },
            "description": "Nombre max de résultats"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Nombre de résultats à ignorer",
              "default": 0,
              "title": "Offset"
            },
            "description": "Nombre de résultats à ignorer"
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/train/intent/versions": {
      "get": {
        "tags": [
          "Intent Training v1"
        ],
        "summary": "List Intent Versions",
        "description": "Versions d'intention valides + pointeur actif : {total, items, active}.\n\n``items`` : liste de noms de versions (tri DESC). ``active`` : version\nrésolue par défaut (pointeur active.json, sinon la plus récente valide) ;\n``None`` si aucun modèle n'existe (repli de règles du classifieur).",
        "operationId": "list_intent_versions_api_v1_train_intent_versions_get",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Intent Versions Api V1 Train Intent Versions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/train/intent/activate": {
      "post": {
        "tags": [
          "Intent Training v1"
        ],
        "summary": "Activate Intent Version Endpoint",
        "description": "Pointe active.json sur une version existante (422 si inconnue).\n\nLe classifieur en mémoire n'est PAS rechargé ici : l'IHM chaîne ensuite\nPOST /classifiers/intent/reload pour appliquer la nouvelle version.",
        "operationId": "activate_intent_version_endpoint_api_v1_train_intent_activate_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntentActivateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Activate Intent Version Endpoint Api V1 Train Intent Activate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IntentActivateRequest": {
        "properties": {
          "version": {
            "type": "string",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "version"
        ],
        "title": "IntentActivateRequest",
        "description": "Corps de POST /train/intent/activate : version à activer.",
        "examples": [
          {
            "version": "20260905T120000Z"
          }
        ]
      },
      "IntentTrainRequest": {
        "properties": {
          "dataset_path": {
            "type": "string",
            "title": "Dataset Path",
            "default": "data/intent_dataset.jsonl"
          },
          "base_model": {
            "type": "string",
            "title": "Base Model",
            "default": "sentence-transformers/paraphrase-multilingual-MiniLM-L12-v2"
          },
          "base_model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Model Version"
          },
          "epochs": {
            "type": "integer",
            "title": "Epochs",
            "default": 3
          },
          "batch_size": {
            "type": "integer",
            "title": "Batch Size",
            "default": 32
          },
          "learning_rate": {
            "type": "number",
            "title": "Learning Rate",
            "default": 2e-05
          },
          "max_length": {
            "type": "integer",
            "title": "Max Length",
            "default": 64
          },
          "test_size": {
            "type": "number",
            "title": "Test Size",
            "default": 0.1
          },
          "quantize_int8": {
            "type": "boolean",
            "title": "Quantize Int8",
            "default": false
          },
          "activate": {
            "type": "boolean",
            "title": "Activate",
            "default": false
          }
        },
        "type": "object",
        "title": "IntentTrainRequest",
        "description": "Paramètres de POST /train/intent (entraînement du classifieur d'intention).\n\nReprise des options du script ``scripts/train_intent.py`` : dataset JSONL\n``{\"text\", \"label\"}`` (labels fixés à chat/action), encodeur HF,\nquantification INT8 optionnelle et activation de la version produite.\n``base_model_version`` déclenche le continual training depuis une version\nexistante de ``experiments/intent_models`` (validée tôt par la route).",
        "examples": [
          {
            "activate": false,
            "base_model": "sentence-transformers/paraphrase-multilingual-MiniLM-L12-v2",
            "batch_size": 32,
            "dataset_path": "data/intent_dataset.jsonl",
            "epochs": 3,
            "learning_rate": 2e-05,
            "max_length": 64,
            "quantize_int8": false,
            "test_size": 0.1
          }
        ]
      },
      "JobListResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/TrainJob"
            },
            "type": "array",
            "title": "Items"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "total",
          "items",
          "limit",
          "offset"
        ],
        "title": "JobListResponse",
        "description": "RÃ©ponse paginÃ©e pour GET /train/jobs."
      },
      "JobStatus": {
        "type": "string",
        "enum": [
          "pending",
          "running",
          "completed",
          "failed",
          "cancelled"
        ],
        "title": "JobStatus"
      },
...
      "TrainJob": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "step": {
            "type": "string",
            "title": "Step",
            "default": "queued"
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "default": "train"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "model_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Path"
          },
          "regression": {
            "type": "boolean",
            "title": "Regression",
            "default": false
          },
          "regression_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Regression Detail"
          },
          "progress": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status"
        ],
        "title": "TrainJob"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}